fix(tests): fix broken tests due the refactor
