+++ src/main/factories/controllers/convert-number-to-full-text-controller-factory.ts
@@ -1,10 +1,10 @@
-import { Controller } from '../../../presentation/protocols';
 import { ConvertNumberToFullTextController } from '../../../presentation/controllers/convert-number-to-full-text-controller';
-import { makeConvertNumberToFullTextUsecase } from '../usecases/convert-number-to-full-text-usecase-factory';
+import { Controller } from '../../../presentation/protocols';
+import { makeConvertNumberToFullText } from '../usecases/convert-number-to-full-text-usecase-factory';
 import { makeNumberValidator } from '../validators/make-number-validator-factory';
 
 export const makeConvertNumberToFullTextController = (): Controller => {
   const numberValidator = makeNumberValidator();
-  const convertor = makeConvertNumberToFullTextUsecase();
+  const convertor = makeConvertNumberToFullText();
   return new ConvertNumberToFullTextController(numberValidator, convertor);
 };
